rename ist to layers
Update README.md
