+++ vptq/__init__.py
@@ -4,4 +4,4 @@
 # --------------------------------------------------------------------------
 
 __version__ = "0.0.1"
-from .ist.model_base import AutoModelForCausalLM as AutoModelForCausalLM
+from .layers.model_base import AutoModelForCausalLM as AutoModelForCausalLM
