+++ vptq/__main__.py
@@ -7,7 +7,7 @@ import argparse
 
 import transformers
 
-from .ist.model_base import AutoModelForCausalLM as VQAutoModelQuantization
+from .layers.model_base import AutoModelForCausalLM as VQAutoModelQuantization
 
 
 def define_basic_args():
